ihm_fultang-main/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-js-asset==3.1.2
django-mptt==0.16.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Faker==35.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyMuPDF==1.25.3
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
